+++ Source/Game/Components/SpreadDetectComponent.h
@@ -2,10 +2,10 @@
 #include <glm/vec2.hpp>
 #include <queue>
 #include "Component.h"
-#include "../../Constants/SpreadConstants.h"
+#include "../World/SpreadKey.h"
 
-const uint8_t MAX_DETECT = 16;
-const uint8_t MAX_ADD_DISTANCE = 2;
+const int MAX_DETECT = 16;
+const int MAX_ADD_DISTANCE = 2;
 const glm::ivec2 NO_ADD = glm::ivec2(-1024, -1024);
 
 class SpreadDetectComponent : public Component {
@@ -22,7 +22,7 @@ public:
     SpreadDetectComponent & operator=(const SpreadDetectComponent &) = delete;
 
     static constexpr std::string GetName() { return "spread_detect"; }
-    static constexpr uint8_t GetID() { return 8; }
+    static constexpr int GetID() { return 8; }
 
     void Load(nlohmann::json& data, EntityID entity) {
     }
